backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI DriveThru Backend API"
authors = [
    {name = "yithril",email = "yithril@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pytest (>=8.4.2,<9.0.0)",
    "llm-guard (>=0.3.16,<0.4.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "aiofiles>=23.2.0",
    "aioboto3>=12.0.0",
    "httpx>=0.25.0",
    "boto3>=1.34.0",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "fastapi-nextauth-jwt (>=2.1.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"